cocoa_common: remove deprecated VOCTRLs/VO_EVENTs
See commit 35b73f9412e29a and the following as an example how this would
have to be done if done properly.

Since I'm unable to test on OSX, and nobody is interested in fixing this
code (including myself, actually), just remove the deprecated
definitions to make sure the code still builds. This will break runtime
switching of fullscreen, ontop, border. (The way the minimized state is
reported was also deprecated, but commit 981c921c827 already broke it
anyway.)